+++ src/OpenTelemetry/CHANGELOG.md
@@ -9,6 +9,10 @@ please check the latest changes
 
 ## Unreleased
 
+## 1.1.0-rc1
+
+Released 2021-Jun-25
+
 * Moved `IDeferredTracerProviderBuilder` to API library.
   ([#2058](https://github.com/open-telemetry/opentelemetry-dotnet/pull/2100))